vocab.py => vocab module
