+++ kospeech/checkpoint/checkpoint.py
@@ -12,7 +12,7 @@ import torch.nn as nn
 from kospeech.utils import logger
 from kospeech.data.data_loader import SpectrogramDataset
 from kospeech.models.las.model import ListenAttendSpell
-from kospeech.optim.optimizer import Optimizer
+from kospeech.optim import Optimizer
 
 
 class Checkpoint(object):
