+++ kospeech/vocabs/__init__.py
@@ -1,0 +1,20 @@
+# -*- coding: utf-8 -*-
+# Soohwan Kim, Seyoung Bae, Cheolhwang Won.
+# @ArXiv : KoSpeech: Open-Source Toolkit for End-to-End Korean Speech Recognition
+# This source code is licensed under the Apache 2.0 License license found in the
+# LICENSE file in the root directory of this source tree.
+
+
+class Vocabulary(object):
+    """
+    Note:
+        Do not use this class directly, use one of the sub classes.
+    """
+    def __init__(self, *args, **kwargs):
+        self.sos_id = None
+        self.eos_id = None
+        self.pad_id = None
+        self.blank_id = None
+
+    def label_to_string(self, labels):
+        raise NotImplementedError
